+++ .eslintrc.js
@@ -15,9 +15,6 @@ module.exports = {
 					'error',
 					{
 						enforceBuildableLibDependency: true,
-						// These packages may not need to build because they are not published.
-						// All packages in @guardian are published to npm so must be built.
-						allow: ['@csnx/**'],
 						depConstraints: [
 							{
 								sourceTag: '*',